+++ Assets/Code/Word.cpp
@@ -12,13 +12,13 @@ Word::~Word()
 
 
 void Word::init(int x, int y, int scaleX, int scaleY, int separationX, string word){
-	ElementHUD::init(x, y, NULL, false, 0, 0, 0, 0, scaleX, scaleY);
+	ElementHUD::init(x, y, NULL, false, 0, 0, 0, 0, static_cast<float>(scaleX), static_cast<float>(scaleY));
 
 	int _x = x;
 	int _y = y;
 	for (int i = 0; i < word.length(); i++){
 		lettersNums.push_back(new ABCsAlphaNum());
-		lettersNums.back()->init(_x, _y, scaleX, scaleY, word[i]);
+		lettersNums.back()->init(_x, _y, static_cast<float>(scaleX), static_cast<float>(scaleY), word[i]);
 		_x += separationX;
 	}
 }
